Many bug fixes on CPU, Bootloader 100% functional. GPU is working (finally). Still having trouble to run many games. Have yet to implement all the rest.

+++ include/cpu/Z80.hpp
@@ -6,8 +6,8 @@
 namespace CPU {
 
   typedef struct s_timer_clock {
-    uint16_t t;
-    uint16_t m;
+    int t;
+    int m;
   } CLOCK;
   
   class Z80 {
